requirements.txt
sqlalchemy == 1.3.20
discord.py == 1.5.0
python-dotenv == 0.15
mysql-connector-python == 8.0.22
lark-parser == 0.11.1
